File uploading&downloading

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const mongoose = require('mongoose')
 
+const fileHelper = require('../util/file');
+
 const {
     validationResult
 } = require('express-validator/check');
@@ -20,9 +22,28 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+
+    console.log(image);
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            path: '/admin/add-product',
+            pageTitle: 'Add Product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'upload a valid image png/jpg/jpeg',
+            validationErrors: []
+        })
+    }
+
     const errors = validationResult(req)
 
     if (!errors.isEmpty()) {
@@ -41,6 +62,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: error.array()
         })
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
@@ -97,7 +120,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
 
@@ -114,7 +137,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -130,7 +152,10 @@ exports.postEditProduct = (req, res, next) => {
                 return res.redirect('/');
             }
             product.title = updatedTitle;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             product.price = updatedPrice;
             product.description = updatedDescription;
             return product.save().then(result => {
@@ -168,9 +193,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({
-            _id: prodId,
-            userId: req.user._id
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('product not found'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({
+                _id: prodId,
+                userId: req.user._id
+            })
         })
         .then(() => {
             console.log('product deleted')
